requirements.txt
astroid==1.3.6
Django==1.8
django-autofixture==0.10.0
django-bootstrap3==5.4.0
django-braces==1.8.0
django-debug-toolbar==1.3.0
django-vanilla-views==1.0.3
flake8==2.4.0
logilab-common==0.63.2
psycopg2==2.6
pylint==1.4.3
selenium==2.45.0
six==1.9.0
sqlparse==0.1.15

